moved rpc calls to rpc directory
