+++ snek_sploit/lib/rpc/__init__.py
@@ -1,0 +1,2 @@
+from snek_sploit.lib.rpc.auth import RPCAuth
+from snek_sploit.lib.rpc.console import RPCConsole, ConsoleInfo
